+++ sparql_service/src/srdf_data/rdf_data_error.rs
@@ -1,0 +1,48 @@
+use std::io;
+
+use oxigraph::{
+    sparql::{EvaluationError, SparqlSyntaxError},
+    store::StorageError,
+};
+use thiserror::Error;
+
+use srdf::{SRDFGraphError, SRDFSparqlError};
+
+#[derive(Debug, Error)]
+pub enum RdfDataError {
+    #[error(transparent)]
+    SRDFSparqlError {
+        #[from]
+        err: SRDFSparqlError,
+    },
+
+    #[error(transparent)]
+    SRDFGraphError {
+        #[from]
+        err: SRDFGraphError,
+    },
+
+    #[error(transparent)]
+    IOError {
+        #[from]
+        err: io::Error,
+    },
+
+    #[error(transparent)]
+    StorageError {
+        #[from]
+        err: StorageError,
+    },
+
+    #[error(transparent)]
+    SparqlParseError {
+        #[from]
+        err: SparqlSyntaxError,
+    },
+
+    #[error(transparent)]
+    SparqlEvaluationError {
+        #[from]
+        err: EvaluationError,
+    },
+}